requirements.txt
## GUI
streamlit==1.36.0

## DB Clients
qdrant_client==1.10.1

## LlamaIndex
llama-index==0.10.52
llama-parse==0.4.4
llama-index-readers-file==0.1.27
llama-index-embeddings-openai==0.1.10
llama-index-llms-openai==0.1.21
llama-index-llms-langchain==0.1.4
llama-index-postprocessor-cohere-rerank==0.1.7
llama-index-callbacks-langfuse==0.1.4
llama-index-vector-stores-qdrant==0.2.14
llama-index-storage-docstore-mongodb==0.1.3

## LangChain
langchain==0.1.20
langchainhub==0.1.20

## Observability
langfuse==2.36.2

## Cloud Services
google-cloud-secret-manager==2.20.0
google-crc32c==1.5.0